platform/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
azure-ai-projects>=2.0.0b3
azure-identity
python-dotenv
httpx
